requirements.txt
flask
requests
gunicorn

google-cloud-vision
openai
